+++ source/tasks/Generate_Wiki_Content.build.ps1
@@ -18,6 +18,11 @@
         The path to the source folder name. Defaults to the same path where the
         module manifest is found.
 
+    .PARAMETER WikiSourceFolderName
+        The name of the folder that contain the source markdown files (e.g. 'Home.md')
+        to publish to the wiki. The name should be relative to the SourcePath.
+        Defaults to 'WikiSource'.
+
     .PARAMETER BuildInfo
         The build info object from ModuleBuilder. Defaults to an empty hashtable.
 
@@ -77,18 +82,32 @@ param
     ),
 
     [Parameter()]
+    [System.String]
+    $WikiSourceFolderName = (property WikiSourceFolderName 'WikiSource'),
+
+    [Parameter()]
     [System.Collections.Hashtable]
     $BuildInfo = (property BuildInfo @{ })
 )
 
 # Synopsis: This task generates wiki documentation for the DSC resources.
 task Generate_Wiki_Content {
+    #Import-Module -Name 'DscResource.DocGenerator' -Force
+
     if (-not (Split-Path -IsAbsolute $OutputDirectory))
     {
         $OutputDirectory = Join-Path -Path $ProjectPath -ChildPath $OutputDirectory
     }
 
-    $wikiOutputPath = Join-Path -Path $OutputDirectory -ChildPath "WikiContent"
+    $getBuiltModuleVersionParameters = @{
+        OutputDirectory = $OutputDirectory
+        ProjectName     = $ProjectName
+    }
+
+    $moduleVersion = Get-BuiltModuleVersion @getBuiltModuleVersionParameters
+
+    $wikiOutputPath = Join-Path -Path $OutputDirectory -ChildPath 'WikiContent'
+
     if ((Test-Path -Path $wikiOutputPath) -eq $false)
     {
         $null = New-Item -Path $wikiOutputPath -ItemType Directory
@@ -96,10 +115,36 @@ task Generate_Wiki_Content {
 
     "`tProject Path            = $ProjectPath"
     "`tProject Name            = $ProjectName"
+    "`tModule Version          = $moduleVersion"
     "`tSource Path             = $SourcePath"
     "`tWiki Output Path        = $wikiOutputPath"
+
+    $wikiSourcePath = Join-Path -Path $SourcePath -ChildPath $WikiSourceFolderName
+
+    $wikiSourceExist = Test-Path -Path $wikiSourcePath
+
+    if ($wikiSourceExist)
+    {
+        "`tWiki Source Path        = $wikiSourcePath"
+    }
 
     Write-Build Magenta "Generating Wiki content for all DSC resources based on source."
 
     New-DscResourceWikiPage -ModulePath $SourcePath -OutputPath $wikiOutputPath
+
+    if ($wikiSourceExist)
+    {
+        Write-Build Magenta "Copying Wiki content from the Wiki source folder."
+
+        Copy-Item -Path (Join-Path $wikiSourcePath -ChildPath '*') -Destination $wikiOutputPath -Force
+
+        $homeMarkdownFilePath = Join-Path -Path $wikiOutputPath -ChildPath 'Home.md'
+
+        if (Test-Path -Path $homeMarkdownFilePath)
+        {
+            Write-Build Magenta "Updating module version in Home.md if there are any placeholders found."
+
+            Set-WikiModuleVersion -Path $homeMarkdownFilePath -ModuleVersion $moduleVersion
+        }
+    }
 }